REF: try to separate pyepics PV compat layer
